+++ thumbnailer.h
@@ -4,11 +4,11 @@
 struct Buffer {
     uint8_t* data;
     size_t size;
-    uint32_t width, height;
+    uint64_t width, height;
 };
 
 struct Dims {
-    uint32_t width, height;
+    uint64_t width, height;
 };
 
 // Writes RGBA thumbnail buffer to img
